pyproject.toml
[tool.poetry]
name = "cat-ai-service"
version = "0.1.0"
description = "HTTP API for retrieving descriptions and tags for cat photos from OpenAI"
authors = ["Haiko Schol <hs@haikoschol.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
httpx = "^0.26.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.27.0.post1"
openai = "^1.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
